fix(result): Fix typecheck in `get_or`
If a function failes (raises an exception), the internal result value is
set to `None`. So the type of that value is always `None` on failure, which
breaks the typecheck of `get_or`.

The typecheck now relies on type hints because it can't deduce that
information otherwise.

There is a also a TODO to improve this. See Issue #2 for more information.